Aide à l'intégration dans un nouveau composant
